fix/custom-media-elements-timing: add intermediary promise method for checking if media resembles a media element before invoking mediaSetCallback. Break out isMediaEl into a util. (per PR feedback)
